requirements.txt
customtkinter>=5.2.0
yfinance>=0.2.28
ta>=0.10.2
mplfinance>=0.12.9
matplotlib>=3.7.0
google-generativeai>=0.3.0
